Move person_to_observation function to ehr_union from combine
