+++ data_steward/test/test_data/five_persons/person.csv
@@ -1,6 +1,6 @@
 person_id,gender_concept_id,year_of_birth,month_of_birth,day_of_birth,birth_datetime,race_concept_id,ethnicity_concept_id,location_id,provider_id,care_site_id,person_source_value,gender_source_value,gender_source_concept_id,race_source_value,race_source_concept_id,ethnicity_source_value,ethnicity_source_concept_id
-10,1,1999,10,5,1981-09-15 12:40:13.343317 UTC,10,10,10,10,10,Google Adds,Sales,10,Google Adds,10,St. Paul,10
-8,2,1996,11,11,1999-09-29 05:15:08.467313 UTC,8,8,8,8,8,Word of mouth,Direct,8,Direct,8,Virginia Beach,8
-6,2,2011,8,17,1975-10-04 19:11:36.435910 UTC,6,6,6,6,6,Previous Customer,Add. Words,6,Internet,6,Chicago,6
-9,2,2017,2,7,1971-09-08 07:41:09.037623 UTC,9,9,9,9,9,Sales,Internet,9,Add. Words,9,Anchorage,9
-7,2,1986,6,30,1974-04-05 12:09:51.092328 UTC,7,7,7,7,7,Sales,Add. Words,7,Direct,7,Kansas,7
+20,1,1999,10,5,1981-09-15 12:40:13.343317 UTC,10,10,10,10,10,Google Adds,Sales,10,Google Adds,10,St. Paul,10
+18,2,1996,11,11,1999-09-29 05:15:08.467313 UTC,8,8,8,8,8,Word of mouth,Direct,8,Direct,8,Virginia Beach,8
+16,2,2011,8,17,1975-10-04 19:11:36.435910 UTC,6,6,6,6,6,Previous Customer,Add. Words,6,Internet,6,Chicago,6
+19,2,2017,2,7,1971-09-08 07:41:09.037623 UTC,9,9,9,9,9,Sales,Internet,9,Add. Words,9,Anchorage,9
+17,2,1986,6,30,1974-04-05 12:09:51.092328 UTC,7,7,7,7,7,Sales,Add. Words,7,Direct,7,Kansas,7
