+++ data_steward/test/test_data/pitt_five_person/observation.csv
@@ -1,6 +1,6 @@
 "observation_id","person_id","observation_concept_id","observation_date","observation_datetime","observation_type_concept_id","value_as_number","value_as_string","value_as_concept_id","qualifier_concept_id","unit_concept_id","provider_id","visit_occurrence_id","observation_source_value","observation_source_concept_id","unit_source_value","qualifier_source_value"
-"6","6","6","1996-09-14","1989-10-07 22:13:11.1533266","6","832.66752717675058","GJTXJ2SIOOO92MMRZIG431F8W2AMP8GG8M","6","6","6","6","6","Add. Words","5","Internet","Internet"
-"7","7","7","1990-06-03","1960-01-24 18:00:23.4031621","7","49.08448646268085","20RU9IJZEM71GNQAYFMCE4K1HR5TG781FEJQ8N7LK20MO1UY","7","7","7","7","7","Internet","6","Sales","Add. Words"
-"8","8","8","1959-09-12","1987-04-15 06:06:30.6260548","8","1253.5394752647446","Y","8","8","8","8","8","Sales","7","Sales","Previous Customer"
-"9","9","9","1975-05-09","1982-02-26 12:35:59.6336904","9","1555.1072552591129","1LASYGPDHCPI7D6M1401B81SC4XCF","9","9","9","9","9","Previous Customer","8","Word of mouth","Sales"
-"10","10","10","1963-01-31","1983-05-13 15:31:52.3905384","10","579.16600097863284","7UIP","10","10","10","10","10","Word of mouth","9","Previous Customer","Word of mouth"
+"16","6","6","1996-09-14","1989-10-07 22:13:11.1533266","6","832.66752717675058","GJTXJ2SIOOO92MMRZIG431F8W2AMP8GG8M","6","6","6","6","6","Add. Words","5","Internet","Internet"
+"17","7","7","1990-06-03","1960-01-24 18:00:23.4031621","7","49.08448646268085","20RU9IJZEM71GNQAYFMCE4K1HR5TG781FEJQ8N7LK20MO1UY","7","7","7","7","7","Internet","6","Sales","Add. Words"
+"18","8","8","1959-09-12","1987-04-15 06:06:30.6260548","8","1253.5394752647446","Y","8","8","8","8","8","Sales","7","Sales","Previous Customer"
+"19","9","9","1975-05-09","1982-02-26 12:35:59.6336904","9","1555.1072552591129","1LASYGPDHCPI7D6M1401B81SC4XCF","9","9","9","9","9","Previous Customer","8","Word of mouth","Sales"
+"20","10","10","1963-01-31","1983-05-13 15:31:52.3905384","10","579.16600097863284","7UIP","10","10","10","10","10","Word of mouth","9","Previous Customer","Word of mouth"
